+++ tests/suites/tenant/diagnostics/tabs/schema.test.ts
@@ -1,0 +1,28 @@
+import {expect, test} from '@playwright/test';
+
+import {dsVslotsSchema, tenantName} from '../../../../utils/constants';
+import {TenantPage} from '../../TenantPage';
+import {Diagnostics, DiagnosticsTab} from '../Diagnostics';
+
+test.describe('Diagnostics tab', async () => {
+    test('Primary keys header is visible in Schema tab', async ({page}) => {
+        const pageQueryParams = {
+            schema: dsVslotsSchema,
+            database: tenantName,
+            tenantPage: 'diagnostics',
+        };
+        const tenantPage = new TenantPage(page);
+        await tenantPage.goto(pageQueryParams);
+
+        const objectSummary = new Diagnostics(page);
+
+        await objectSummary.clickTab(DiagnosticsTab.Schema);
+        await expect(objectSummary.isSchemaViewerVisible()).resolves.toBe(true);
+
+        await expect(objectSummary.getPrimaryKeys()).resolves.toEqual([
+            'NodeId',
+            'PDiskId',
+            'VSlotId',
+        ]);
+    });
+});
